Hints are now much, much better. I’ll have to discuss them with you all, but primarily, there’s no more need for a hint view controller in the storyboard. Rather, it’s merely dependent on whether or not you have run the “addNewHint…” function. Also, that is much easier to set up for because the BGKVPasswordControls keep track of the number of attempts done on them so far.
Mm, in actuality now that I think about it, that won’t work as desired without caching. Like, you’d have to attempt the password three times IN ONE GO. But, that is solved by caching, which is as simple as setting ‘cacheKey’ to something unique! So that’s fine!

Currently there are automatically some dummy hints.