Ohhh god. I’ve been working all night. There has been an awful lot done. Well, good news the cutscenes appear before the levels now, so there isn’t a glimpse of the level. (Except if you go to the level directly from the level select screen. Then there is.)
Also, I have edited literally every storyboard so this is a major conflict point. That’s because I got rid of all the levelContainers and StoryboardLinks (in the storyboard. They still exist in code.)

The cutscenes code has ended up getting changed a lot too, just in the course of using it.